requirements.txt
Pillow==6.1.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
blis==0.4.1
Bottleneck==1.3.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
cymem==2.0.3
fastai==1.0.60
fastprogress==0.2.2
h11==0.9.0
httptools==0.0.13
idna==2.8
importlib-metadata==1.4.0
kiwisolver==1.1.0
matplotlib==3.1.2
more-itertools==8.1.0
multidict==4.7.4
murmurhash==1.0.2
numexpr==2.7.1
numpy==1.18.1
nvidia-ml-py3==7.352.0
packaging==20.0
pandas==0.25.3
plac==1.1.3
preshed==3.0.2
pynvx==1.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
scipy==1.4.1
six==1.13.0
soupsieve==1.9.5
spacy==2.2.3
srsly==1.0.1
starlette==0.13.0
thinc==7.3.1
torch==1.3.1
torchvision==0.4.2
tqdm==4.41.1
urllib3==1.25.7
uvicorn==0.11.1
uvloop==0.14.0
wasabi==0.6.0
websockets==8.1
yarl==1.4.2
zipp==0.6.0
